Avoid defining make_splits using S3 methods
This is a hack to avoid an issue when building rsample with the
webr-repo scripts. For some reason, the build fails when make_splits
is defined using S3 methods, but not when make_splits is defined as
a normal function.

With this, the ability to use make_splits with a data.frame is no longer
available.